- removes redundant font for size cache lookups
- also caches editor setting
- usies a faster hash method on the FontForSizeAdvanced cache (its the same function its just not converting the key to a variant before hashing)
- ShapedTextDataAdvanced::valid changed to an SafeFlag to avoid locking a mutex, as its read Very often

in total from my benchmark it takes ~39% of the time to render the same text in a Label (~45% in CodeEdit, the main script editor in godot)
all of these optimizations came from profiling a build using `target=editor optimize=speed_trace` they all gave at least a 10% improvement to speed at the time of profiling
now when profiled, the actual hotpath goes to freetype rendering and shaping and such

benchmark: [TextServerTest.zip](https://github.com/user-attachments/files/15508175/TextServerTest.zip)
benchmark compile flags: `scons target=template_release production=yes optimize=speed lto=full tests=yes deprecated=no scu_build=yes module_text_server_adv_enabled=yes module_text_server_fb_enabled=no`
benchmark run command: `./godot/bin/godot.linuxbsd.template_release.x86_64 --disable-vsync ./TextServerTest/project.godot`

results before:
```
Godot Engine v4.3.beta.custom_build.6b38be699 (2024-05-30 20:38:39 UTC) - https://godotengine.org
Vulkan 1.3.277 - Forward+ - Using Device #0: NVIDIA - NVIDIA GeForce GTX 1080 Ti

18.1370287230521ms avg draw time (99 samples)
18.1469531974407ms avg draw time (198 samples)
18.1223431018868ms avg draw time (297 samples)
18.159361198695ms avg draw time (396 samples)
18.1944639995845ms avg draw time (495 samples)
18.1336226286712ms avg draw time (594 samples)
18.14051692792ms avg draw time (693 samples)
18.1234557219226ms avg draw time (792 samples)
18.0797271856957ms avg draw time (891 samples)
18.0556610377148ms avg draw time (990 samples)
18.0405357542993ms avg draw time (1089 samples)
```
after:
```
Godot Engine v4.3.beta.custom_build.02f8b6443 (2024-05-30 22:54:24 UTC) - https://godotengine.org
Vulkan 1.3.277 - Forward+ - Using Device #0: NVIDIA - NVIDIA GeForce GTX 1080 Ti

7.03977334378946ms avg draw time (99 samples)
7.00848030321526ms avg draw time (198 samples)
7.01666921878905ms avg draw time (297 samples)
7.02174504597982ms avg draw time (396 samples)
7.03076449307528ms avg draw time (495 samples)
7.036976942711ms avg draw time (594 samples)
7.03303741686272ms avg draw time (693 samples)
7.02608956231011ms avg draw time (792 samples)
7.01870688150494ms avg draw time (891 samples)
7.01467798213766ms avg draw time (990 samples)
7.0109612576123ms avg draw time (1089 samples)
```

the last line of each is the most important, its the average of all frame times (minus some longer startup frames)

the same improvements can be made to TextServerFallback