@akien-mga 2024-08-27
Needs a rebase, which should show a compilation error (tested locally):
```
In file included from ./core/string/char_utils.h:34,
                 from ./core/string/ustring.h:36,
                 from modules/text_server_adv/script_iterator.h:46,
                 from modules/text_server_adv/text_server_adv.h:39,
                 from modules/text_server_adv/text_server_adv.cpp:31:
modules/text_server_adv/text_server_adv.cpp: In member function 'virtual PackedInt32Array TextServerAdvanced::_font_get_supported_glyphs(const RID&) const':
modules/text_server_adv/text_server_adv.cpp:3603: error: no matching function for call to 'TextServerAdvanced::_ensure_cache_for_size(TextServerAdvanced::FontAdvanced*&, Vector2i) const'
 3603 |                 ERR_FAIL_COND_V(!_ensure_cache_for_size(fd, fd->msdf ? Vector2i(fd->msdf_source_size, 0) : Vector2i(16, 0)), PackedInt32Array());
./core/typedefs.h:272:41: note: in definition of macro 'unlikely'
  272 | #define unlikely(x) __builtin_expect(!!(x), 0)
      |                                         ^
modules/text_server_adv/text_server_adv.cpp:3603: note: in expansion of macro 'ERR_FAIL_COND_V'
 3603 |                 ERR_FAIL_COND_V(!_ensure_cache_for_size(fd, fd->msdf ? Vector2i(fd->msdf_source_size, 0) : Vector2i(16, 0)), PackedInt32Array());
modules/text_server_adv/text_server_adv.cpp:1360: note: candidate: 'bool TextServerAdvanced::_ensure_cache_for_size(FontAdvanced*, const Vector2i&, FontForSizeAdvanced*&) const'
 1360 | _FORCE_INLINE_ bool TextServerAdvanced::_ensure_cache_for_size(FontAdvanced *p_font_data, const Vector2i &p_size, FontForSizeAdvanced *&r_cache_for_size) const {
modules/text_server_adv/text_server_adv.cpp:1360: note:   candidate expects 3 arguments, 2 provided
```